Unit Tests: fix tests for IE
IE has some strange behavior. It modifies the value of the inline styling, as fully explained on https://github.com/joequery/Stupid-Table-Plugin/issues/128

Best thing to avoid this silly problem is to simply compare the inline styling before any JS is run "on the relevant table element". And then compare this value with the one after the relevant JS is run.