Ensure string functions are handling strings
If you tag a column as being sorted by "string" or "string-ins" and have
a cell with a numeric value, you will receive the error "a.localeCompare
is not a function" or "a.toLocaleLowerCase is not a function." This is
because it's trying to perform the function on a number. Converted
everything toString() to prevent this.